dev insert,patch,delete wishlist with error status

+++ app-dai/src/main/java/com/example/appdai/controller/PcController.java
@@ -99,21 +99,43 @@ public class PcController {
             }
         });
 
+        // INSERT: insert un élément de la liste de souhaits
+        app.patch("/api/users/{userId}/add/wishlist/{photocardId}", ctx -> {
+            Integer userId = Integer.parseInt(ctx.pathParam("userId"));
+            Integer photocardId = Integer.parseInt(ctx.pathParam("photocardId"));
+            Boolean have = Boolean.parseBoolean(ctx.pathParam("have"));
+            if (userId != null || photocardId != null) {
+                cardService.removeFromWishlist(userId, photocardId,have);
+                ctx.status(204).result("Photocard ajoutée à la wishlist");
+            }else{
+                ctx.status(404).result("Le user n'existe pas ou la carte n'existe pas");
+            }
+            
+        });
+
         // PATCH: Modifier un élément de la liste de souhaits (having)
-        app.patch("/api/users/{userId}/wishlist/{photocardId}", ctx -> {
+        app.patch("/api/users/{userId}/update/wishlist/{photocardId}", ctx -> {
             Integer userId = Integer.parseInt(ctx.pathParam("userId"));
             Integer photocardId = Integer.parseInt(ctx.pathParam("photocardId"));
             Photocard updatedPhotocard = ctx.bodyAsClass(Photocard.class);
-            cardService.updateWishlist(userId, photocardId, updatedPhotocard);
-            ctx.status(200).result("Photocard mise à jour");
+            if (userId != null || photocardId != null) {
+                cardService.updateWishlist(userId, photocardId, updatedPhotocard);
+                ctx.status(200).result("Photocard mise à jour");
+            }else{
+                ctx.status(404).result("Le user n'existe pas ou la carte n'existe pas");
+            }
         });
 
         // DELETE: Supprimer un élément de la liste de souhaits
         app.delete("/api/users/{userId}/wishlist/{photocardId}", ctx -> {
             Integer userId = Integer.parseInt(ctx.pathParam("userId"));
             Integer photocardId = Integer.parseInt(ctx.pathParam("photocardId"));
-            cardService.removeFromWishlist(userId, photocardId);
-            ctx.status(204).result("Photocard supprimée de la wishlist");
+            if (userId != null || photocardId != null) {
+                cardService.removeFromWishlist(userId, photocardId);
+                ctx.status(204).result("Photocard supprimée de la wishlist");
+            }else{
+                ctx.status(404).result("Le user n'existe pas ou la carte n'existe pas");
+            }
         });
 
     }
